Rename ServerLoop -> Server
This is the baseline server class, which uses GreenSocket to accept connections.
It does not depend directly on pyuv and is compatible with any loop backend
including the epoll fallback backend.

There are other ways to implement the Server class by depending on pyuv and a
normal socket to accept(), but in my tests, this mechanism did not show a
significant performance increase and only led to increased library complexity.